Restaurant-DonPepe-main/RestaurantSoft-main/restaurant-backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
Pillow
python-dotenv
